chore: rewrite examples/lnclient scripts in TypeScript
